+++ apps/studio/src/routes/editor/TopBar/index.tsx
@@ -6,7 +6,7 @@ import { Tooltip, TooltipContent, TooltipTrigger } from '@onlook/ui/tooltip';
 import { observer } from 'mobx-react-lite';
 import ModeToggle from './ModeToggle';
 import OpenCode from './OpenCode';
-import PricingButton from './Profile/PricingButton';
+import PricingButton from './Profile';
 import ProjectBreadcrumb from './ProjectSelect';
 import ShareProject from './ShareProject';
 import ZoomControls from './ZoomControls';
@@ -83,7 +83,6 @@ const EditorTopBar = observer(() => {
                 <ZoomControls />
                 <OpenCode />
                 <ShareProject />
-                {/* <ProfileButton /> */}
                 <PricingButton />
             </div>
         </div>
